Remove digest package's dependency on external sha implementation
The change relies on a refactor of the upstream resumable sha256/sha512 package
that opts to register implementations with the standard library. This allows
the resumable support to be detected where it matters, avoiding unnecessary and
complex code. It also ensures that consumers of the digest package don't need
to depend on the forked sha implementations.

We also get an optimization with this change. If the size of data written to a
digester is the same as the file size, we check to see if the digest has been
verified. This works if the blob is written and committed in a single request.

+++ Godeps/_workspace/src/github.com/stevvooe/resumable/sha512/resume.go
@@ -6,8 +6,8 @@ import (
 )
 
 // Len returns the number of bytes which have been written to the digest.
-func (d *digest) Len() uint64 {
-	return d.len
+func (d *digest) Len() int64 {
+	return int64(d.len)
 }
 
 // State returns a snapshot of the state of the digest.
